+++ build.rs
@@ -1,19 +1,31 @@
-const COMMANDS: &[&str] = &[];
-const PERMISSIONS: &'static str = r#"
+const PERMISSIONS_FOR_ANDROID_9_OR_LOWER: &'static str = r#"
 <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" android:maxSdkVersion="28" />
 <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" android:maxSdkVersion="28" />
 "#;
 
+const PERMISSIONS_FOR_ANDROID_10_OR_LOWER: &'static str = r#"
+<uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" android:maxSdkVersion="29" />
+<uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" android:maxSdkVersion="29" />
+"#;
+
 fn main() {
-	tauri_plugin::Builder::new(COMMANDS)
+	tauri_plugin::Builder::new(&[])
 		.android_path("android")
 		.build();
 
-	if std::env::var("CARGO_FEATURE_LEGACY_STORAGE_PERMISSION").is_ok() {
+	if std::env::var("CARGO_FEATURE_LEGACY_STORAGE_PERMISSION_INCLUDE_ANDROID_10").is_ok() {
 		tauri_plugin::mobile::update_android_manifest(
 			"ANDROID FS PLUGIN",
 			"manifest",
-			PERMISSIONS.trim().to_string()
+			PERMISSIONS_FOR_ANDROID_10_OR_LOWER.trim().to_string()
+		)
+		.expect("failed to rewrite AndroidManifest.xml");
+	}
+	else if std::env::var("CARGO_FEATURE_LEGACY_STORAGE_PERMISSION").is_ok() {
+		tauri_plugin::mobile::update_android_manifest(
+			"ANDROID FS PLUGIN",
+			"manifest",
+			PERMISSIONS_FOR_ANDROID_9_OR_LOWER.trim().to_string()
 		)
 		.expect("failed to rewrite AndroidManifest.xml");
 	}